6th_Nov/backend/requirements.txt
fastapi
uvicorn
requests
yfinance
python-dotenv
langchain
